# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "alwism.com - API Documentation",
        "description": "This is the API Gateway for [www.alwism.com](https://www.alwism.com).  You can find out more about alwism at [https://www.alwism.com/sitemechanism](https://www.alwism.com/sitemechanism) or visit to github repository [github](http://url.cc/).",
        "contact": {
            "name": "Contact the Developer",
            "url": "https://www.alwism.com/contactme"
        },
        "version": "1.0.3"
    },
    "servers": [
        {
            "url": "api/v1"
        }
    ],
    "paths": {
...
        "/api/admin/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get All Projects",
                "operationId": "bd52029fb3c372f16fabc66541a454db",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/GetProjects"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/project/{slug}": {
            "get": {
                "tags": [
                    "Non Auth"
                ],
                "summary": "Get Project by slug",
                "operationId": "7f5f615517bcb9bb443b7925f346b0f8",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/GetbySlug"
                    }
                }
            }
        },
        "/api/admin/project/store": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Store a newly created project",
                "operationId": "0391df77f100b9e9cf58003e063a202e",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Project"
                },
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "400": {
                        "description": "bad request: Some/all required fields are empty or Duplicate Slug Url"
                    },
                    "406": {
                        "description": "Not Acceptable: Do not change user Id. for the demonstration please use 3"
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/project/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get Project by ID",
                "operationId": "1838ecd9ba43e933ba2a5d7eeb6fdd72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/GetbySlug"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update the specified project in storage.",
                "operationId": "3c1330bf05095cada7be449bd8855664",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Project"
                },
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "400": {
                        "description": "bad request: Some/all required fields are empty or Duplicate Slug Url"
                    },
                    "406": {
                        "description": "Not Acceptable: Do not change user Id, for the demonstration please use 3,"
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/project/delete/{id}": {
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Remove the specified project from server.",
                "operationId": "db861c9690a19faba8a104132d1d809a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/urlcheck": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Check Project Slug Url",
                "operationId": "273d3433eb5a1abbf02c4da8d272eb77",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "url": {
                                        "description": "Test Url",
                                        "type": "string",
                                        "example": "abc-traders"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete todo success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "responses": {
...
            "GetProjects": {
                "description": "success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "current_page": {
                                    "type": "number",
                                    "example": "1"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                },
                                "first_page_url": {
                                    "type": "string",
                                    "example": "url/api/projects?page=1"
                                },
                                "from": {
                                    "type": "number",
                                    "example": "1"
                                },
                                "last_page": {
                                    "type": "number",
                                    "example": "1"
                                },
                                "last_page_url": {
                                    "type": "string",
                                    "example": "url/api/projects?page=3"
                                },
                                "links": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "url": {
                                                "type": "string",
                                                "example": "url/api/projects?page=1"
                                            },
                                            "label": {
                                                "type": "string",
                                                "example": "&laquo; Previous"
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": "false"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "next_page_url": {
                                    "type": "string",
                                    "example": "null"
                                },
                                "path": {
                                    "type": "string",
                                    "example": "url/api/projects"
                                },
                                "per_page": {
                                    "type": "number",
                                    "example": "15"
                                },
                                "prev_page_url": {
                                    "type": "string",
                                    "example": "url/api/projects?page=1"
                                },
                                "to": {
                                    "type": "number",
                                    "example": "2"
                                },
                                "total": {
                                    "type": "number",
                                    "example": "2"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "GetbySlug": {
                "description": "success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": "true"
                                },
                                "slug": {
                                    "type": "string",
                                    "example": "abc-car-traders"
                                },
                                "data": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "userId": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Ebook LTD ASP.NET MVC 6"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "abc-car-traders"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "path/image.png"
                                        },
                                        "description": {
                                            "type": "text",
                                            "example": "Project Description"
                                        },
                                        "seoTitle": {
                                            "type": "string",
                                            "example": "ABC Car Traders"
                                        },
                                        "seoDescription": {
                                            "type": "string",
                                            "example": "SEO Description"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "publish"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Go Lang"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "categories": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Go Lang"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "requestBodies": {
            "Project": {
                "description": "Project Request Body",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "$ref": "#/components/schemas/Project"
                        }
                    },
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Project"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}